% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_analyse_record.R
\name{acdc_access}
\alias{acdc_access}
\alias{acdc_access_dat}
\alias{acdc_access_timesteps}
\alias{acdc_access_maps}
\title{Short-cuts to elements of an \code{"acdc_record"} object}
\usage{
acdc_access_dat(record)

acdc_access_timesteps(record)

acdc_access_maps(
  record,
  type = c("map_timestep", "map_cumulative"),
  select = NULL
)
}
\arguments{
\item{record}{An \code{\link[flapper]{acdc_record-class}} object.}

\item{type}{For \code{\link[flapper]{acdc_access_maps}}, \code{type} is a character that specifies whether or not to access time step-specific maps (\code{type = "map_timestep"}) or  cumulative maps (\code{type = "map_cumulative"}).}

\item{select}{(optional) For \code{\link[flapper]{acdc_access_maps}}, \code{select} is an integer vector that defines the cumulative time steps (i.e., accounting for both primary (acoustic) and secondary (archival) time steps) for which maps are required.}
}
\value{
\itemize{
  \item \code{\link[flapper]{acdc_access_dat}} returns a single dataframe for all time steps.
  \item \code{\link[flapper]{acdc_access_timesteps}} returns an integer that defines the total number of time steps.
  \item \code{\link[flapper]{acdc_access_maps}} returns a single list of time-step specific or cumulative maps for specified or all time steps.
}
}
\description{
These functions provide short-cuts to elements of an \code{\link[flapper]{acdc_record-class}} object.
}
\details{
\itemize{
  \item \code{\link[flapper]{acdc_access_dat}} accesses the \code{record$dat} elements of an \code{\link[flapper]{acdc_record-class}} object for all time steps.
  \item \code{\link[flapper]{acdc_access_timesteps}} accesses the total number of time steps stored in an \code{\link[flapper]{acdc_record-class}} object, accounting for both primary (acoustic) and secondary (archival) time steps.
  \item \code{\link[flapper]{acdc_access_maps}} accesses all, or specified, maps from the \code{record$spatial} elements of an \code{\link[flapper]{acdc_record-class}} object.
}
}
\examples{
#### Example (1): acdc_access_dat()
acdc_access_dat(acdc_simplify(dat_acdc))

#### Example (2): acdc_access_timesteps()
acdc_access_timesteps(acdc_simplify(dat_acdc))

#### Example (3): acdc_access_maps()
acdc_access_maps(acdc_simplify(dat_acdc))

}
\author{
Edward Lavender
}
