% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_support.R
\name{acdc_helper_access}
\alias{acdc_helper_access}
\alias{acdc_helper_access_timesteps}
\title{Short-cuts to elements of an \code{".acdc"} object}
\usage{
acdc_helper_access_timesteps(x)
}
\arguments{
\item{x}{An \code{\link[flapper]{.acdc-class}} object.}
}
\description{
These functions provide short-cuts to elements of a \code{\link[flapper]{.acdc-class}} object.
}
\details{
\itemize{
  \item \code{\link[flapper]{acdc_helper_access_timesteps}} gets the total number of time steps from an AC* algorithm implementation.
}
}
\examples{
#### Example (1): acdc_helper_access_timesteps()
acdc_helper_access_timesteps(acdc_simplify(dat_acdc))

}
\author{
Edward Lavender
}
