% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_support.R
\name{acdc_helper_list_record_for_pf}
\alias{acdc_helper_list_record_for_pf}
\title{List `record' files from an AC/DC algorithm}
\usage{
acdc_helper_list_record_for_pf(root, type = c("acs", "dc"), ...)
}
\arguments{
\item{root}{A string that defines the directory in which files are loaded.}

\item{type}{A character that defines the source of the files (\code{type = "acs"} refers to an AC* algorithm and \code{type = "dc"} refers to the DC algorithm).}

\item{...}{Additional arguments passed to \code{\link[base]{list.files}} (excluding \code{full.names}).}
}
\value{
The function returns an ordered list of file paths.
}
\description{
This function creates an ordered list of `record' files derived from the AC/DC/ACDC algorithm.
}
\details{
This function requires the \code{\link[stringr]{stringr}} package.
}
\examples{
#### Example (1): Example with the AC algorithm
# Define a directory in which to save files
root <- paste0(tempdir(), "/ac/")
dir.create(root)
# Implement the AC algorithm for some example time series
acc <- dat_acoustics[dat_acoustics$individual_id == 25, ][1:5, ]
out_ac <- ac(acoustics = acc,
             step = 120,
             bathy = dat_gebco,
             detection_range = 425,
             acc_centroids = dat_centroids,
             mobility = 250,
             write_record_spatial_for_pf = list(filename = root))
# List the files for pf()
files <- acdc_helper_list_record_for_pf(root, type = "acs", pattern = "*.grd")
utils::head(files)
# Implement pf() using files (not shown).

#### Example (2): Example with the DC algorithm
# Define a directory in which to save files
root <- paste0(tempdir(), "/dc/")
dir.create(root)
# Implement the DC algorithm for some example time series
depth <- dat_archival[dat_archival$individual_id == 25, ][1:5, ]
out_dc <- dc(archival = depth,
             bathy = dat_gebco,
             write_record_spatial_for_pf = list(filename = root))
# List the files for pf()
files <- acdc_helper_list_record_for_pf(root, type = "dc", pattern = "*.grd")
utils::head(files)
# Implement pf() using files (not shown).

}
\author{
Edward Lavender
}
