% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc.R
\name{acdc_setup_n_centroids}
\alias{acdc_setup_n_centroids}
\title{Suggest the number of centroids for the ACDC algorithm}
\usage{
acdc_setup_n_centroids(
  detections,
  moorings,
  mobility,
  double = TRUE,
  hist = TRUE,
  ...
)
}
\arguments{
\item{detections}{A POSIXct vector of time stamps when detections were made (for a particular individual).}

\item{moorings}{A dataframe that defines passive acoustic telemetry receiver locations and deployment periods. This must contain the following columns: 'receiver_id', a unique identifier of each receiver; 'receiver_lat' and 'receiver_long', the latitude and longitude of each receiver in decimal degrees; and 'receiver_start_date' and 'receiver_end_date', the start and end time of each receiver's deployment period (see \code{\link[flapper]{dat_moorings}} for an example).}

\item{mobility}{A number that defines the distance (m) that an individual could move in the time period between archival observations (see also \code{\link[flapper]{acdc_setup_centroids}}).}

\item{double}{A logical variable that defines whether or not to double the minimum number of centroids given by approach two (see Details).}

\item{hist}{A logical value that defines whether or not to plot the distribution of gaps across the detection time series as a histogram.}

\item{...}{Additional arguments passed to \code{\link[prettyGraphics]{pretty_hist}}.}
}
\value{
The function returns an integer vector with the upper and lower suggested value for the number of centroids from methods (1) and (2). The parameters used to generate these suggestions (i.e., \code{detections}, \code{moorings}, \code{mobility} and \code{double}) are also included in a 'param' attribute.
}
\description{
The ACDC algorithm requires a list of acoustic centroids from \code{\link[flapper]{acdc_setup_centroids}}. The number of centroids that is required depends on (a) the duration between detections; (b) the distances among receivers; (c) the area of interest; and (d) other considerations such as the presence of a long tail of archival observations after the final acoustic detection. This function implements two approaches to make suggestions as to suitable numbers of centroids for the ACDC algorithm (see Details). This requires a time series of detections, the deployment details of passive acoustic telemetry receivers and a mobility parameter. Given these inputs, the function returns a suggested upper and lower bound for the number of centroids.
}
\details{
\subsection{Method (1)}{The first approach provides a reasonable upper bound for the number of centroids. This approach is based on the gaps between detections. During this time, acoustic centroids increase in size, reflecting the increasing uncertainty in the location of an individual, until the half way point between detections. At this point, uncertainty in the individual's location is maximised and the acoustic centroids reach their maximum size. Thereafter, uncertainty in the individual's location and the size of the acoustic centroids decrease towards the receiver at which the individual was next detected. Under this perspective, the minimum number of centroids is the number required such that the centroids continue to increase in size until the halfway point between the two most temporally distance detections. More specifically, assuming that archival time steps are 2-minutes in duration, the minimum number of centroids is one quarter of the duration (minutes) of the longest gap between acoustic detections. To implement this approach, a time series of \code{detections} (for the individual and time period for which the ACDC algorithm will be implemented) needs to be provided. This approach provides a sensible upper bound for the number of centroids, since it covers the full area in which movement is possible. However, even with modest gaps between detections and a modest movement capacity, this may suggest a very large area, and it may take a long time to compute the requisite number of centroids using \code{\link[flapper]{acdc_setup_centroids}}. Moreover, in practice, the area of interest may be smaller.}

\subsection{Method (2)}{The second approach provides a reasonable lower bound for the number of centroids. This approach is based on the locations of receivers and the assumption that, when an individual is detected by two different receivers, at the halfway point between detections, its potential location is described by the intersection of the acoustic centroid around the receiver at which it was previously detected and the centroid around the receiver at which it is next detected (evaluated at the halfway point between detections). Under this perspective, the minimum centroid size is half of the distance between the furthest two receivers that were operational at the same time. (This could be restricted to the subset of receivers at which the individual was detected sequentially but this is not implemented.) In practice, it is advisable to double this minimum number to ensure a reasonable degree of overlap between the centroids of the two receivers. To implement this approach, a dataframe that contains receiver locations and deployment periods (\code{moorings}) must be supplied, along with the \code{mobility} parameter that describes how far an individual could move in any archival time step. This approach provides a sensible lower bound for the number of centroids that is array-specific. However, it does not account for the full span of movements that are possible under the ACDC model.}

\subsection{Other considerations}{In practice, the most appropriate number of centroids is likely to be a compromise between these minimum and maximum values that depends on other considerations, particularly the spatial scale of the study, the effect of boundaries on the final centroid size and, perhaps in some cases, the geographical range of the species. For acoustic time series that are followed by a long tail of archival observations after final acoustic detection, both approaches may underestimate the number of centroids required to capture the increasing uncertainty in the individual's location over this time. However, in this case, it is advisable to use the depth-contour (DC) algorithm (see \code{\link[flapper]{dc}}) at some point after the last acoustic detection since the influence of that observation on putative patterns of space use will decay through time and the DC algorithm is more computationally efficient.}
}
\examples{
n_timesteps <-
  acdc_setup_n_centroids(dat_acoustics$timestamp[dat_acoustics$individual_id == 25],
                         dat_moorings,
                         mobility = 200,
                         double = TRUE)
utils::str(n_timesteps)

}
\seealso{
This function is designed to facilitate an informed choice for the 'n_timesteps' argument in \code{\link[flapper]{acdc_setup_centroids}}. This underpins the ACDC algorithm, which is implemented by \code{\link[flapper]{acdc}}.
}
\author{
Edward Lavender
}
