% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_setup.R
\name{acs_setup_n_centroids}
\alias{acs_setup_n_centroids}
\title{Suggest the number of centroids for the AC* algorithm(s)}
\usage{
acs_setup_n_centroids(
  detections,
  step,
  hist = TRUE,
  ...,
  moorings,
  mobility,
  double = TRUE,
  boundaries
)
}
\arguments{
\item{detections}{For Method One, \code{detections} is a POSIXct vector of time stamps when detections were made (for a particular individual).}

\item{step}{For Method One, \code{step} is the duration (s) between sequential time steps inbetween acoustic detections. For the ACDC algorithm, this is the duration between sequential archival observations.}

\item{hist, ...}{For Method One, \code{hist} is a logical value that defines whether or not to plot the distribution of gaps across the detection time series as a histogram and \code{...} includes any additional arguments passed to \code{\link[prettyGraphics]{pretty_hist}}.}

\item{moorings}{For Method Two, \code{moorings} is a dataframe that defines passive acoustic telemetry receiver locations and deployment periods. This must contain the following columns: `receiver_id', a unique identifier of each receiver; `receiver_lat' and `receiver_long', the latitude and longitude of each receiver in decimal degrees; and `receiver_start_date' and `receiver_end_date', the start and end time of each receiver's deployment period (see \code{\link[flapper]{dat_moorings}} for an example).}

\item{mobility}{For Methods Two and Three, \code{mobility} is a number that defines the distance (m) that an individual could move in the time period in the time steps between acoustic observations (see also \code{\link[flapper]{acs_setup_centroids}}).}

\item{double}{For Method Two, \code{double} is a logical variable that defines whether or not to double the minimum number of centroids (see Details).}

\item{boundaries}{For Method Three, a Spatial* or Raster* object that defines the study site, from which area boundaries can be extracted (via \code{\link[raster]{extent}}).}
}
\value{
The function prints a table with the suggested number of centroids and the maximum radius size from each method. Invisibly, the function returns an integer vector with three suggested values for the number of centroids from methods (1), (2) and (3). The parameters used to generate these suggestions (i.e., \code{detections}, \code{moorings}, \code{mobility}, \code{double} and \code{boundaries}) are also included in a `param' attribute.
}
\description{
The acoustic-centroid (AC) and acoustic-centroid depth-contour (ACDC) algorithms require a list of acoustic centroids from \code{\link[flapper]{acs_setup_centroids}}. The number of centroids that is required depends on (a) the duration between detections; (b) the distances among receivers; (c) the area of interest; and (d) other considerations such as the spatial extent of the study area. This function implements three methods to facilitate a sensible choice of the number of centroids for the AC* algorithm(s) (see Details). This ensures that the number of centroids is sufficient to cover requirements but not so large that the centroids become slow to compute and unwieldy. The function requires a time series of detections, the time step length, the deployment details of passive acoustic telemetry receivers, a mobility parameter and a Spatial* or Raster* layer that defines the boundaries of the study site. Given these inputs, the function returns suggested bounds for the number of centroids.
}
\details{
\subsection{Method (1)}{The first method provides a reasonable upper bound for the number of centroids. This method is based on the gaps between detections. During this time, acoustic centroids increase in size, reflecting the increasing uncertainty in the location of an individual, until the half way point between detections. At this point, uncertainty in the individual's location is maximised and the acoustic centroids reach their maximum size. Thereafter, uncertainty in the individual's location and the size of the acoustic centroids decrease towards the receiver at which the individual was next detected. Under this perspective, the minimum number of centroids is the number required such that the centroids continue to increase in size until the halfway point between the two most temporally distance detections. For example, if the time steps between acoustic detections are two-minutes in duration, the minimum number of centroids is one quarter of the duration (minutes) of the longest gap between acoustic detections. To implement this method, a time series of \code{detections} (for the individual and time period for which the AC* algorithm(s) will be implemented) and the time \code{step} length need to be provided. This method provides a sensible upper bound for the number of centroids, since it allows centroids to continue to expand to their maximum possible size, given the data and the assumptions of the algorithm. However, even with modest gaps between detections and a modest movement capacity, this may suggest a very large area, and it may take a long time to compute the requisite number of centroids using \code{\link[flapper]{acs_setup_centroids}}. Moreover, in practice, the area of interest may be smaller.}

\subsection{Method (2)}{The second method provides a reasonable lower bound for the number of centroids. This method is based on the locations of receivers and the assumption that, when an individual is detected by two different receivers, at the halfway point between detections, its potential location is described by the intersection of the acoustic centroid around the receiver at which it was previously detected and the centroid around the receiver at which it is next detected (evaluated at the halfway point between detections). Under this perspective, the minimum centroid size is half of the distance between the furthest two receivers that were operational at the same time and the minimum centroid number is this size over the animal's mobility†. (This could be restricted to the subset of receivers at which the individual was detected sequentially but this is not implemented.) In practice, it is advisable to double this minimum number to ensure a reasonable degree of overlap between the centroids of the two receivers. To implement this method, a dataframe that contains receiver locations and deployment periods (\code{moorings}) must be supplied, along with the \code{mobility} parameter that describes how far an individual could move in any archival time step. This method provides a sensible lower bound for the number of centroids that is array-specific. However, it does not account for the full span of movements that are possible under the AC* algorithm(s).}

\subsection{Method (3)}{The final method provides an alternative suggestion for the number of centroids based on the minimum number that is required for the largest centroid around each receiver to fill the study area. This method requires the locations of receivers and a Spatial* or Raster* layer that defines the study area, the boundaries of which are approximated as a bounding box with four vertices and transformed to degrees latitude/longitude. For each receiver, the function calculates the distance to each of the four boundary coordinates. The longest distance between a receiver and a boundary of the study area is used to define the number of centroids required for the largest centroid to fill the study area†. In many settings, this method may be most suitable because it accounts for the full span of movements that are possible under the AC* algorithm(s) while focusing on a specific area.}

\subsection{Other considerations}{In practice, the most appropriate number of centroids is likely to be a compromise between these minimum and maximum values that depends on other considerations, particularly the spatial scale of the study, the effect of boundaries on the final centroid size and, perhaps in some cases, the geographical range of the species. For acoustic time series that are followed by a long tail of archival observations after final acoustic detection, both methods may underestimate the number of centroids required to capture the increasing uncertainty in the individual's location over this time. However, in this case, it is advisable to use the depth-contour (DC) algorithm (see \code{\link[flapper]{dc}}) at some point after the last acoustic detection since the influence of that observation on putative patterns of space use will decay through time and the DC algorithm is more computationally efficient. If in doubt, opt to make more, rather than fewer centroids: if there are too few centroids, the AC* algorithm(s) can fail to converge if centroids around receivers with sequential detections do not overlap.}

†Both Methods Two and Three assume that the detection range is the same as the \code{mobility}. In practice, this is unlikely to be the case, but if the detection range is even approximately equal to \code{mobility} then the difference is negligible.
}
\examples{
n_timesteps <-
  acs_setup_n_centroids(detections = dat_acoustics$timestamp[dat_acoustics$individual_id == 25],
                         step = 120,
                         moorings = dat_moorings,
                         mobility = 200,
                         double = TRUE,
                         boundaries = dat_gebco)
utils::str(n_timesteps)

}
\seealso{
This function is designed to facilitate an informed choice for the `n_timesteps' argument in \code{\link[flapper]{acs_setup_centroids}}. \code{\link[flapper]{acs_setup_mobility}} can also guide the implementation of this function. This underpins the AC and ACDC algorithms, which are implemented by \code{\link[flapper]{ac}} and \code{\link[flapper]{acdc}}.
}
\author{
Edward Lavender
}
