% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_receiver_id.R
\name{add_receiver_id}
\alias{add_receiver_id}
\title{Add unique receiver IDs to passive acoustic telemetry timeseries}
\usage{
add_receiver_id(acoustics, moorings)
}
\arguments{
\item{acoustics}{A dataframe which comprises passive acoustic telemetry detection timeseries. This must contain two named columns: 'timestamp', a POSIXct vector which defines the time of each detection; and 'receiver', a vector which defines the receiver at which the individual was detected. The column 'receiver' should also be found in \code{moorings} (see below).}

\item{moorings}{A dataframe which contains passive acoustic telemetry receiver metadata. This must contain four named columns: 'receiver', as above for \code{moorings}; 'receiver_id', a unique identifier for each receiver deployment; 'start_date', a POSIXct vector which defines the time of each receiver deployment; and 'end_date', a POSIXct vector which defines the end of each receiver deployment. If objects of class Date are provided for 'start_date' and 'end_date', these are coerced to POSIXct objects with a warning so that the timing of receiver detections and deployment is comparable (i.e., of the same object type).}
}
\value{
The function returns a vector of receiver IDs, as defined in the \code{moorings$receiver_id} column, which correspond to each detection in the \code{acoustics} dataframe.
}
\description{
This function extracts the unique receiver deployment IDs (e.g., 1,...n) from a dataframe containing receiver attributes (e.g. unique receiver deployment IDs, receiver codes and locations), termed \code{moorings}, that correspond to passive acoustic telemetry (PAT) receiver codes in detection timeseries, termed \code{acoustics}. This is especially useful if the same receiver has been deployed multiple times (e.g. in different locations) over the course of the study. In this scenario, the receiver code in the PAT detection timeseries does not uniquely identify the unique receiver deployment, which means that receiver codes in PAT data and metadata cannot simply be matched: instead, both receiver code(s) and the time of detection(s) need to be included in the matching procedure. In this case, \code{add_receiver_ids()} examines each receiver with detections in \code{acoustics}. If that receiver was redeployed, the function examines the time of each detection and returns the ID of the receiver that recorded that detection (given the timing of receiver deployment). If each receiver was only deployed once, this function simply matches receiver codes in \code{moorings} and receiver codes in \code{acoustics} and returns the receiver IDs (which, in this case, may be the same as the receiver codes) in \code{moorings} that correspond to each receiver code in \code{acoustics}.
}
\details{
The function implements \code{\link[data.table]{foverlaps}} to overlap the timing of detections with the timing of receiver deployments to account for receiver deployment in the assignment of receiver IDs.
}
\examples{

}
\author{
Edward Lavender
}
