% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_support.R
\name{check_class_.acdc}
\alias{check_class_.acdc}
\title{Check an object is of class \code{".acdc"}}
\usage{
check_class_.acdc(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{An object.}

\item{arg}{A character that defines the name of a function argument.}
}
\value{
If \code{x} is not a \code{\link[flapper]{.acdc-class}} object, the function returns a helpful error message.
}
\description{
This function checks whether or not an object is of class \code{\link[flapper]{.acdc-class}}.
}
\author{
Edward Lavender
}
\keyword{internal}
