% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_support.R
\name{check_class_acdc_record}
\alias{check_class_acdc_record}
\title{Check an object is of class \code{"acdc_record"}}
\usage{
check_class_acdc_record(x, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{An object.}

\item{arg}{A character that defines the name of a function argument.}
}
\value{
If \code{x} is not an \code{\link[flapper]{acdc_record-class}} object, the function returns a helpful error message.
}
\description{
This function checks whether or not an object is of class \code{\link[flapper]{acdc_record-class}}.
}
\author{
Edward Lavender
}
\keyword{internal}
