% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compact}
\alias{compact}
\title{Compact a list}
\source{
This function is derived from the \code{plyr::compact()} function. The function is defined separately in \code{\link[flapper]{flapper}} to reduce reliance on non-default packages.
}
\usage{
compact(l)
}
\arguments{
\item{l}{A list.}
}
\description{
Remove all \code{NULL} entries from a list.
}
\keyword{internal}
