% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{crop_from_click}
\alias{crop_from_click}
\title{Interactively crop a \code{\link[raster]{raster}}}
\usage{
crop_from_click(x, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link[raster]{raster}}.}

\item{plot}{A logical input that defines whether or not to plot \code{x} before and after cropping.}

\item{...}{Additional arguments passed to \code{\link[prettyGraphics]{pretty_map}} to customise plots.}
}
\value{
The function returns a \code{\link[raster]{raster}} and, if \code{plot = TRUE}, a plot of the area before/after cropping.
}
\description{
Interactively crop a \code{\link[raster]{raster}} to a boundary box specified by mouse clicks.
}
\details{
The function is implemented as follows:
\enumerate{
  \item The function plots the supplied raster (\code{x}), if \code{plot = TRUE};
  \item Boundary coordinates delineating the area to which \code{x} should be cropped are defined by mouse clicks on the plot;
  \item The plotted raster is cropped using the minimum and maximum x and y coordinates of mouse clicks, via \code{\link[raster]{crop}};
  \item The cropped raster is plotted, if \code{plot = TRUE};
  \item The cropped raster is returned.
}
}
\examples{
if(interactive()) crop_from_click(dat_gebco)
}
\author{
Edward Lavender
}
