% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_acdc}
\alias{dat_acdc}
\title{Example ACDC algorithm output}
\format{
A named list with 4 elements:
\describe{
  \item{.acdc}{A list of results from calls to \code{\link[flapper]{.acdc}}, with one element per chunk.}
  \item{ts_by_chunk}{A list of time series, with one element per chunk.}
  \item{time}{A dataframe that defines the times of sequential stages in the algorithm's progression.}
  \item{args}{A named list of user inputs that record the parameters used to generate the outputs.}
}
}
\usage{
dat_acdc
}
\description{
An object of class \code{\link[flapper]{acdc-class}} from \code{\link[flapper]{acdc}}, created by a (2-hour) chunk-wise implementation of this function. This is included principally to streamline function examples.
}
\keyword{datasets}
