% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_coast}
\alias{dat_coast}
\title{The coastline around the MEFS Firth of Lorn acoustic array}
\format{
A SpatialPolygonsDataFrame (see \code{\link[sp]{SpatialPolygonsDataFrame-class}}).
}
\source{
https://biogeo.ucdavis.edu/data/gadm3.6/Rsp/gadm36_GBR_0_sp.rds
}
\usage{
dat_coast
}
\description{
A SpatialPolygonsDataFrame delineating the coastline around a subset of acoustic receivers set up by the Movement Ecology of Flapper Skate (MEFS) project in the Firth of Lorn, off the west coast of Scotland.
}
\keyword{datasets}
