% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_containers}
\alias{dat_containers}
\title{Example detection containers from \code{\link[flapper]{acs_setup_containers}}}
\format{
A list with 57 elements, one for each receiver from 1:max(\code{\link[flapper]{dat_moorings}$receiver_id}). Each element contains a SpatialPolygonsDataFrame with the detection container for that receiver, under a detection range of 425 m and within the boundaries defined by \code{\link[flapper]{dat_coast}} and \code{\link[flapper]{dat_gebco}}.
}
\usage{
dat_containers
}
\description{
A list of detection containers created by \code{\link[flapper]{acs_setup_containers}}. This is included principally to streamline function examples.
}
\keyword{datasets}
