% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_dc}
\alias{dat_dc}
\title{Example DC algorithm output}
\format{
A named list with 4 elements:
\describe{
  \item{archive}{A list of results from internal function calls that contain essential results, with one element per chunk.}
  \item{ts_by_chunk}{A list of time series, with one element per chunk.}
  \item{time}{A dataframe that defines the times of sequential stages in the algorithm's progression.}
  \item{args}{A named list of user inputs that record the parameters used to generate the outputs.}
}
}
\usage{
dat_dc
}
\description{
An object of class \code{\link[flapper]{acdc_archive-class}} from \code{\link[flapper]{dc}} created by a step-wise implementation of this function. This is included principally to streamline function examples.
}
\seealso{
See \code{\link[flapper]{acdc_archive-class}} for further information on this S3 class.
}
\keyword{datasets}
