% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_dcpf}
\alias{dat_dcpf}
\title{Example DCPF algorithm output}
\format{
A named list, with two elements, which records the parameters used in the call to \code{\link[flapper]{dcpf}} and the reconstructed paths:
\describe{
  \item{args}{A named list that records the function arguments used to generate outputs.}
  \item{dcpf}{A dataframe with seven columns and 180 observations that records the reconstructed paths. This contains the following columns:}
  \itemize{
    \item{path_id}{ An integer that uniquely defines each path.}
    \item{timestep}{ An integer that defines each time step.}
    \item{cell_id}{ An integer that defines the cells ID of the surface raster (over which paths were reconstructed) retained by the algorithm.}
    \item{cell_x}{ A double that defines the cell x coordinate.}
    \item{cell_y}{ A double that defines the cell y coordinate.}
    \item{cell_z}{ A double that defines the value of the surface raster in each cell.}
    \item{cell_pr}{ A double that defines the probability of movement into each cell, given immediately preceding sampled locations (or the origin).}
  }
}
}
\usage{
dat_dcpf
}
\description{
A named list of results from an implementation of the depth-contour particle filtering (DCPF) algorithm (see \code{\link[flapper]{dcpf}}). The algorithm was implemented for a series of depth observations collected from flapper skate (\emph{Dipturus intermedius}) off the west coast of Scotland, where the bathymetry is described by \code{\link[flapper]{dat_gebco}}. The outputs include algorithm parameters and a dataframe comprising some of the possible paths over the seabed reconstructed from example depth observations. This is included principally to streamline function examples.
}
\keyword{datasets}
