% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_dcpf_paths}
\alias{dat_dcpf_paths}
\title{Example DCPF algorithm output: reconstructed paths}
\format{
A dataframe with seven columns and 180 observations that records the reconstructed paths. This contains the following columns:
\describe{
    \item{path_id}{ An integer that uniquely defines each path.}
    \item{timestep}{ An integer that defines each time step.}
    \item{cell_id}{ An integer that defines the cells ID of the surface raster (over which paths were reconstructed) retained by the algorithm.}
    \item{cell_x}{ A double that defines the cell x coordinate.}
    \item{cell_y}{ A double that defines the cell y coordinate.}
    \item{cell_z}{ A double that defines the value of the surface raster in each cell.}
    \item{cell_pr}{ A double that defines the probability of movement into each cell, given immediately preceding sampled locations (or the origin).}
  }
}
\usage{
dat_dcpf_paths
}
\description{
An example \code{\link[flapper]{dcpf-class}} object. This comprises a dataframe of movement paths reconstructed by an implementation of the depth-contour particle filtering (DCPF) algorithm (see \code{\link[flapper]{dcpf}}). The algorithm was implemented for a series of depth observations collected from flapper skate (\emph{Dipturus intermedius}) off the west coast of Scotland, where the bathymetry is described by \code{\link[flapper]{dat_gebco}}. The resultant particle histories (\code{\link[flapper]{dat_dcpf_histories}}) were processed into this set of movement paths by \code{\link[flapper]{dcpf_simplify}}. This dataset is included principally to streamline function examples.
}
\keyword{datasets}
