% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_ids}
\alias{dat_ids}
\title{Example tagged individuals dataset}
\format{
A dataframe with 3 observations and 6 variables:
\describe{
  \item{individual_id}{A number which distinguishes each unique individual.}
  \item{transmitter_id}{A character which defines the acoustic transmitter code.}
  \item{archival_id}{A character which defines the archival tag code.}
  \item{tag_start_date}{A date which defines the date of tag deployment.}
  \item{tag_long}{A number which defines the longitude (decimal degrees) of tag deployment.}
  \item{tag_lat}{A number which defines the latitude (decimal degrees) of tag deployment.}
}
}
\source{
Data were collected by Marine Scotland Science and Scottish Natural Heritage. Data were processed by Edward Lavender.
}
\usage{
dat_ids
}
\description{
A dataset containing the characteristics of a sample of flapper skate (\emph{Dipturus intermedius}) that were tagged as part of the Movement Ecology of Flapper Skate (MEFS) project. Data are arranged by \code{individual_id} (see below).
}
\keyword{datasets}
