% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf-classes.R
\docType{package}
\name{dcpf-class}
\alias{dcpf-class}
\title{"dcpf" class}
\value{
A dataframe with that records the reconstructed paths. This includes a unique identifier for each path, the time step, the location (cell ID and three-dimensional coordinates) on \code{bathy} and the probability associated with that cell, given movement from the previous cell, in the following columns:
\describe{
    \item{path_id}{ An integer that uniquely defines each path.}
    \item{timestep}{ An integer that defines each time step. If an origin is supplied to \code{\link[flapper]{dcpf}} and \code{add_origin = TRUE} in \code{\link[flapper]{dcpf_simplify}}, then time step 0 refers to the origin. Later time steps refer to sequential depth observations.}
    \item{cell_id}{ An integer that defines the cells ID of the surface raster (over which paths were reconstructed) retained by the algorithm.}
    \item{cell_x}{ A double that defines the cell x coordinate.}
    \item{cell_y}{ A double that defines the cell y coordinate.}
    \item{cell_z}{ A double that defines the value of the surface raster in each cell.}
    \item{cell_pr}{ A double that defines the probability of movement into each cell, given immediately preceding sampled locations (or the origin).}
Rows are ordered by path and then time step.
  }
}
\description{
An S3 class that defines the object returned by \code{\link[flapper]{dcpf_simplify}}, following \code{\link[flapper]{dcpf}}.
}
\seealso{
\code{\link[flapper]{dat_dcpf_paths}} are derived from \code{\link[flapper]{.dcpf-class}} objects via \code{\link[flapper]{dcpf_simplify}}. \code{\link[flapper]{dcpf_simplify}}provides an example.
}
\author{
Edward Lavender
}
