% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf.R
\name{dcpf_plot_2d}
\alias{dcpf_plot_2d}
\title{Map two-dimensional paths from the DCPF algorithm}
\usage{
dcpf_plot_2d(paths, bathy, add_paths = list(), prompt = FALSE, ...)
}
\arguments{
\item{paths}{A dataframe containing reconstructed movement path(s) from \code{\link[flapper]{dcpf}}. At a minimum, this should contain a unique identifier for each path (named `path_id') and the x and y coordinates that define the trajectory of each path (`cell_x' and `cell_y').}

\item{bathy}{A \code{\link[raster]{raster}} of the bathymetry over which movement was reconstructed.}

\item{add_paths}{A named list, passed to \code{\link[prettyGraphics]{add_sp_path}}, to customise the appearance of the paths.}

\item{prompt}{A logical input that defines whether or not plot each path on a separate plot, sequentially, with a pause between plots (\code{prompt = TRUE}), or all paths on a single plot (\code{prompt = FALSE}).}

\item{...}{Additional arguments, passed to \code{\link[prettyGraphics]{pretty_map}}, for plot customisation.}
}
\value{
The function maps the trajectories of reconstructed paths across the bathymetry surface, returning a single map if \code{prompt = FALSE} or one map for each path if \code{prompt = TRUE}.
}
\description{
This function is a simple wrapper for \code{\link[prettyGraphics]{pretty_map}} that maps the paths reconstructed by the DCPF algorithm over a surface.
}
\examples{
#### Implement dcpf() algorithm
# Here, we extract the necessary information from pre-defined outputs for speed
paths    <- dat_dcpf$dcpf
bathy    <- dat_dcpf$args$bathy

#### Example (1): The default implementation
dcpf_plot_2d(paths, bathy)

#### Example (2): Plot customisation options
# Customise the appearance of the path(s)
dcpf_plot_2d(paths, bathy,
             add_paths = list(length = 0.075, col = viridis::viridis(100)))
# Pass arguments to prettyGraphics::pretty_map() via ... , e.g.:
dcpf_plot_2d(paths, bathy, xlab = "Easting (UTM)", ylab = "Northing (UTM)")

#### Example (3): Plot individual paths separately
if(interactive()){
  pp <- par(mfrow = c(3, 4))
  dcpf_plot_2d(paths, bathy, add_paths = list(length = 0.01),
               prompt = TRUE, verbose = FALSE)
  par(pp)
}

}
\seealso{
\code{\link[flapper]{dcpf}} implements the DCPF algorithm. \code{\link[flapper]{dcpf_plot_1d}}, \code{\link[flapper]{dcpf_plot_2d}} and \code{\link[flapper]{dcpf_plot_3d}} provide plotting routines. For mapping, it can be useful to interpolate shortest (least-cost) paths between sequential locations via \code{\link[flapper]{lcp_interp}}. \code{\link[flapper]{dcpf_loglik}} calculates the log-likelihood of each path.
}
\author{
Edward Lavender
}
