% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf.R
\name{dcpf_plot_history}
\alias{dcpf_plot_history}
\title{Plot particle histories from the DCPF algorithm}
\usage{
dcpf_plot_history(
  record,
  time_steps = 1:length(history),
  add_bathy = list(),
  add_particles = list(pch = "."),
  forwards = TRUE,
  prompt = TRUE,
  ...
)
}
\arguments{
\item{record}{A \code{\link[flapper]{.dcpf-class}} object from \code{\link[flapper]{dcpf}} that contains particle histories.}

\item{time_steps}{An integer vector that defines the time steps for which to plot particle histories.}

\item{add_bathy}{A named list, passed to \code{\link[prettyGraphics]{pretty_map}}, to customise the appearance of the bathymetry surface on each map.}

\item{add_particles}{A named list, passed to \code{\link[prettyGraphics]{pretty_map}}, to customise the appearance of the particles on each map.}

\item{forwards}{A logical variable that defines whether or not create plots forwards (i.e., from the first to the last \code{time_steps}) or backwards (i.e., from the last to the first \code{time_steps}).}

\item{prompt}{A logical input that defines whether or not to pause between plots (\code{prompt = TRUE}).}

\item{...}{Plot customisation arguments passed to \code{\link[prettyGraphics]{pretty_map}}.}
}
\value{
The function returns a plot, for each time step, of all the possible locations of the individual, given the \code{archival}, \code{bathy} and \code{calc_depth_error} arguments to code{\link[flapper]{dcpf}} (marked [1]), with sampled locations overlaid.
}
\description{
This function plots the spatiotemporal particle histories from the depth-contour particle filtering (DCPF) algorithm. This produces, for each time step, a map of the individual's possible locations, solely based on its depth (including the bathymetry and the measurement error), with sampled locations overlaid.
}
\examples{
#### Implement dcpf() algorithm
# Here, we use pre-defined outputs for speed

#### Example (1): The default implementation
dcpf_plot_history(dat_dcpf_histories, time_steps = 1)

#### Example (2): Plot customisation options, e.g.:
# Customise bathy via add_bathy()
dcpf_plot_history(dat_dcpf_histories,
                  time_steps = 1,
                  add_bathy = list(col = c(grDevices::topo.colors(2))))
# Customise particles via add_particles
dcpf_plot_history(dat_dcpf_histories,
                  time_steps = 1,
                  add_particles = list(col = "red"))
# Pass other arguments to prettyGraphics::pretty_map() via ...
dcpf_plot_history(dat_dcpf_histories,
                  time_steps = 1,
                  add_polys = list(x = dat_coast, col = "brown"),
                  crop_spatial = TRUE)

#### Example (3): Plot multiple time steps
pp <- graphics::par(mfrow = c(2, 2))
dcpf_plot_history(dat_dcpf_histories, time_steps = 1:4, prompt = FALSE)
graphics::par(pp)

}
\seealso{
\code{\link[flapper]{dcpf}} implements the DCPF algorithm. \code{\link[flapper]{dcpf_simplify}} assembles paths from particle histories. \code{\link[flapper]{dcpf_plot_1d}}, \code{\link[flapper]{dcpf_plot_2d}} and \code{\link[flapper]{dcpf_plot_3d}} provide plotting routines for paths. \code{\link[flapper]{dcpf_loglik}} calculates the log-likelihood of each path.
}
\author{
Edward Lavender
}
