% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf.R
\name{dcpf_simplify}
\alias{dcpf_simplify}
\title{Simplify the outputs of \code{\link[flapper]{dcpf}} for incomplete paths}
\usage{
dcpf_simplify(history, bathy, verbose = TRUE)
}
\arguments{
\item{history}{A named list from an unconverged call to \code{\link[flapper]{dcpf}}.}

\item{bathy}{A \code{\link[raster]{raster}} of the bathymetry for \code{\link[flapper]{dcpf}}.}

\item{verbose}{A logical input that defines whether or not to print messages to the console to monitor function progress.}
}
\value{
The function returns a named list of movement paths, as returned for converged paths by \code{\link[flapper]{dcpf}}.
}
\description{
This function is designed to simplify the list of positions returned by \code{\link[flapper]{dcpf}} for incomplete paths into a set of movement paths. Normally, this routine is implemented by \code{\link[flapper]{dcpf}} internally but if the algorithm fails to converge, an unprocessed list of outputs is returned rather than processed movement paths, in which this function can be used to process the paths.
}
\author{
Edward Lavender
}
