% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_btw_points_3d}
\alias{dist_btw_points_3d}
\title{Calculate the Euclidean distance between points in three-dimensional space.}
\usage{
dist_btw_points_3d(x1, x2, y1, y2, z1, z2)
}
\arguments{
\item{x1}{A number that defines the x-coordinate of the first point.}

\item{x2}{A number that defines the x-coordinate of the second point.}

\item{y1}{A number that defines the y-coordinate of the first point.}

\item{y2}{A number that defines the y-coordinate of the second point.}

\item{z1}{A number that defines the z-coordinate of the first point.}

\item{z2}{A number that defines the z-coordinate of the second point.}
}
\value{
A number that is equal to the Euclidean distance between points in three-dimensional space.
}
\description{
This function returns the Euclidean distance between points in three-dimensional space.
}
\details{
The distance between two points in three dimensional space is given by Pythagoras' Theorem: \eqn{\sqrt{(x_2 - x_1)^2 + (y_2 - y_1)^2 + (z_2 - z_1)^2}}.
}
\examples{
dist_btw_points_3d(1, 2, 1, 2, 1, 2)
}
\author{
Edward Lavender
}
