% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_btw_receivers.R
\name{dist_btw_receivers}
\alias{dist_btw_receivers}
\title{Compute euclidean distances between receivers}
\usage{
dist_btw_receivers(moorings, f = NULL)
}
\arguments{
\item{moorings}{A dataframe which defines the location of each unique receiver combination. This should contain the columns: 'ID', a unique identifier of each receiver, 'lat', the latitude of that receiver in decimal degrees; and 'long', the longitude of that receiver in decimal degrees.}

\item{f}{(optional) A function which is used to process distances before these are returned. For example, it may be useful to round distances to nearest km with \code{f = function(x) round(x, digits = 0)}.}
}
\value{
The function returns a dataframe with columns 'r1', 'r2' and 'dist'. These define the IDs of each combination of receivers and the associated distance between them, in km. Note that the dataframe contains duplicate combinations of receivers (e.g., both r1 = 1 and r2 = 2 and r1 = 2 and r2 = 1).
}
\description{
This function computes linear distances (km) between all combinations of receivers.
}
\examples{

}
\author{
Edward Lavender
}
