% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_classes.R
\docType{package}
\name{.acdc-class}
\alias{.acdc-class}
\title{".acdc" class}
\value{
\subsection{A named list}{An ".acdc" class object is a named list with the following elements: ‘map’, ‘record’, ‘time’, ‘args’, 'chunks' and 'simplify'. The main output element is the ‘map’ RasterLayer that shows where the individual could have spent more or less time over the duration of the movement time series. The ‘record’ element records time-specific maps of the possible locations of the individual, and can be used to plot maps of specific time points or to produce animations (for the time steps specified by the \code{plot} argument). The ‘time’ element is a dataframe that defines the times of sequential stages in the algorithm's progression, providing a record of computation time. The ‘args’ element is a named list of user inputs that record the parameters used to generate the outputs (if \code{keep_args = TRUE}, otherwise the 'args' element is \code{NULL}). The 'chunks' element is a list with chunk-specific information that is generated if the \code{\link[flapper]{acdc}} algorithm is implemented chunk-wise and then simplified via \code{\link[flapper]{acdc_simplify}} with the \code{keep_chunks = TRUE} argument. The \code{simplify} element is a logical value that defines whether or not the object was created from \code{\link[flapper]{acdc}} and \code{\link[flapper]{acdc_simplify}} or \code{\link[flapper]{.acdc}}. Below, more detail about the ‘map’ and ‘record’ elements is provided.}

\subsection{(A) ‘map’}{This is a RasterLayer that defines the locations in which an individual could have been (or was not), given acoustic and archival data. Each cell is a count of the number of times when the acoustic and archival data were compatible with the individual having been in that cell, given the ACDC model. This provides an overall measure of the locations in which an individual could have spent more or less time, but not where it was.}

\subsection{(B) ‘record’}{This is a temporal record of the possible locations of each individual, with one element for each acoustic detection. Each element contains (a) a dataframe that defines the details of each pair of detections (the ‘current’ detection and the ‘next’ detection) and all of the archival observations in-between these detections and (b) a named list of spatial objects, one for each archival time step between the pair of acoustic detections, that records the spatial parameters of the algorithm (e.g., the acoustic centroid) at each time step and the possible location(s) of the individual at that time step based on these parameters.}

\subsection{(i) record$dat}{‘dat’ contains the following columns: ‘timestep_cumulative’ is a cumulative vector that defines the time step of each observation (across all acoustic and archival time steps up to the current time step); ‘timestep_detection’ defines the time step of the detection time series; ‘timestep_archival’ defines the time step of the archival observations between the current and next acoustic detection; ‘receiver_1_id’ and ‘receiver_2_id’ define the receiver at which the individual was detected at the current and next acoustic time steps; ‘receiver_1_timestamp’ and ‘receiver_2_timestamp’ define the time stamps of these detections; ‘time_btw_dets’ provides the time, in minutes, between these acoustic detections, to the nearest two minutes (the assumed duration between archival records); ‘archival_timestamp’ defines the time stamp of these archival observations;  ‘archival_depth’ defines the depth (m) of the individual each archival time step; and ‘centroid_radius’ refers to the size of radius (m) of the acoustic centroid at that time (which depends principally on the time since the last acoustic detection and the time until the next acoustic detection and is a metric of centroid size, bearing-in-mind that the actual centroids can be non-circular).}

\subsection{(ii) record$spatial}{If \code{plot > 0}, then for those time steps specified by \code{plot}, each element in ‘record’ contains a 'spatial' element that is a list, with one element for each archival timestamp between the current and next acoustic detection (defined in ‘record$dat’, see above) that records spatial information pertaining to the possible locations of the individual at that time step. This includes the following elements: ‘centroid’ is a SpatialPolygonsDataFrame that defines the acoustic contour at that time step, which depends on the time since the previous detection, the time until the next detection, the mobility parameter and any spatial masks (i.e., the coastline and specified boundaries); ‘map_timestep’ is a RasterLayer of all the positions that the individual could (cells = 1) or could not (cells = 0) have occupied within that acoustic contour, given its depth and the depth_error, assuming it is on the seabed; and ‘map_cumulative’ is a RasterLayer of the cumulative of the number of times when the acoustic and archival data were compatible with the individual having been in that cell, under the ACDC model, from all previous time steps up to the current time step (i.e., the sum of ‘map_timestep’ across all time steps from the first time step to the current time step). If the receiver at which the individual was detected at the current and next time step differs, after halfway between the archival time steps between these points, the ‘spatial’ list will also contain the following elements: ‘centroid_previous’ is a SpatialPolygonsDataFrame that defines the previous acoustic centroid (i.e., at the ‘current’ receiver); ‘centroid_next’ is a SpatialPolygonsDataFrame that defines the acoustic centroid from the perspective of the ‘next’ receiver at which the individual is detected; 'centroid_overlap' is the overlap between the centroids around the current and next receiver, within which the individual must have been located at the halfway point between detections; and 'centroid_overlap_expanded’ is the expanded version of the centroid overlap that is intersected with the acoustic centroid around the next receiver to determine the individual’s plausible location(s). For \code{plot = 0L} or elements outside of the specified input to \code{plot}, ‘spatial’ is simply an empty list.}
}
\description{
An S3 class object that defines the object returned by the \code{\link[flapper]{acdc}} algorithm, after simplification via \code{\link[flapper]{acdc_simplify}}, or directly from the back-end \code{\link[flapper]{.acdc}} function.
}
\author{
Edward Lavender
}
