% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf-classes.R
\docType{package}
\name{.dcpf-class}
\alias{.dcpf-class}
\title{".dcpf" class}
\value{
This object is a named list, with two elements, which records the parameters used in the call to \code{\link[flapper]{dcpf}} and the particles sampled at each time step:

\describe{
  \item{args}{A named list that records the function arguments used to generate outputs. This is as inputted, but with the `cells_by_time' and `calc_distance_graph' elements added if unsupplied and applicable.}
  \item{history}{A list of dataframes, one for each time step, that record particle samples. Each dataframe comprises \code{n} rows (one for each particle) and the following five columns:}
  \itemize{
    \item{id_previous}{ An integer that uniquely defines a previous location on the bathymetry \code{\link[raster]{raster}} (\code{bathy}). This is absent for the first time step. For subsequent time steps, this is \code{NA} if the fast Euclidean distances method has been used, which does not `remember' previous locations, but specified otherwise.)}
    \item{pr_current}{ A double that defines the movement probabilities associated with previous locations. This is absent for the first time step. For subsequent time steps, this is \code{NA} if the fast Euclidean distances method has been used, which does not `remember' previous locations, but specified otherwise.}
    \item{id_current}{ An integer that uniquely defines each location on the bathymetry \code{\link[raster]{raster}} (\code{bathy}).}
    \item{pr_current} A double that defines the probability of movement into each cell.
    \item{timestep}{ An integer that defines each time step.}
  }
}
}
\description{
An S3 class that defines the object returned by \code{\link[flapper]{dcpf}}.
}
\seealso{
\code{\link[flapper]{dat_dcpf_histories}} provides an example of a \code{\link[flapper]{.dcpf-class}} object. \code{\link[flapper]{dcpf_simplify}} converts \code{\link[flapper]{.dcpf-class}} objects into \code{\link[flapper]{dcpf-class}} objects that comprise the reconstructed movement paths.
}
\author{
Edward Lavender
}
