% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{.planedist2}
\alias{.planedist2}
\title{Calculate a matrix of Euclidean distances between points}
\source{
This function is an internal function in the \code{\link[raster]{raster}} package (https://rdrr.io/rforge/raster/src/R/pointdistance.R). It is defined separately in \code{\link[flapper]{flapper}} for stability.
}
\usage{
.planedist2(p1, p2)
}
\arguments{
\item{p1}{A coordinate matrix.}

\item{p2}{A coordinate matrix.}
}
\description{
Calculate a matrix of Euclidean distances between points
}
\details{
This function assumes that coordinates are in planar space.
}
\keyword{internal}
