% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kud_tools.R
\name{kud_habitat}
\alias{kud_habitat}
\title{Define a `habitat' grid for kernel smoothing}
\usage{
kud_habitat(map, mask = map, mask_inside = FALSE, plot = TRUE, ...)
}
\arguments{
\item{map}{A \code{\link[raster]{raster}} that defines grid properties.}

\item{mask, mask_inside}{Mask options passed to \code{\link[flapper]{mask_io}}.}

\item{plot, ...}{A logical input that defines whether or not to plot the habitat grid, alongside any additional arguments passed to \code{\link[sp]{plot.SpatialPixelsDataFrame}}.}
}
\value{
The function returns a \code{\link[sp]{SpatialPixelsDataFrame}} object that defines `habitat' (1) versus `non-habitat' (0).
}
\description{
This function defines a `habitat' grid for kernel smoothing (e.g., via \code{\link[flapper]{kud_around_coastline}}).
}
\examples{
kud_habitat(map = dat_gebco, mask_inside = FALSE)
kud_habitat(map = dat_gebco, mask = dat_coast, mask_inside = TRUE)
}
\seealso{
\code{\link[flapper]{kud_around_coastline}}
}
\author{
Edward Lavender
}
