% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf-classes.R
\docType{package}
\name{pf-class}
\alias{pf-class}
\title{"pf" class}
\value{
A dataframe with that records the reconstructed paths. This includes a unique identifier for each path, the time step, the location (cell ID and three-dimensional coordinates) on the specified surface (e.g., \code{bathy}) and the probability associated with that cell, given movement from the previous cell, in the following columns:
\describe{
    \item{path_id}{ An integer that uniquely defines each path.}
    \item{timestep}{ An integer that defines each time step. If an origin is supplied to \code{\link[flapper]{pf}} and \code{add_origin = TRUE} in \code{\link[flapper]{pf_simplify}}, then time step 0 refers to the origin. Later time steps refer to sequential depth observations.}
    \item{cell_id}{ An integer that defines the cells ID of the surface \code{\link[raster]{raster}} (over which paths were reconstructed) retained by the algorithm.}
    \item{cell_x}{ A double that defines the cell x coordinate.}
    \item{cell_y}{ A double that defines the cell y coordinate.}
    \item{cell_z}{ A double that defines the value of the surface \code{\link[raster]{raster}} in each cell. If \code{bathy} is unsupplied to \code{\link[flapper]{pf_simplify}} and unavailable in the inputted object, this is NA.}
    \item{cell_pr}{ A double that defines the probability of localisation in each cell, given the `intrinsic' probabilities associated with each cell and the immediately preceding sampled locations (or the origin)'.}
Rows are ordered by path and then time step.
  }
}
\description{
An S3 class that defines the object returned by \code{\link[flapper]{pf_simplify}}, following \code{\link[flapper]{pf}}.
}
\seealso{
\code{\link[flapper]{pf-class}} objects are derived from \code{\link[flapper]{.pf-class}} objects via \code{\link[flapper]{pf_simplify}}. \code{\link[flapper]{dat_dcpf_paths}} provides an example.
}
\author{
Edward Lavender
}
