% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_access.R
\name{pf_access_history}
\alias{pf_access_history}
\title{Access the `history' element of a \code{\link[flapper]{pf_archive-class}} object}
\usage{
pf_access_history(archive, bathy = NULL)
}
\arguments{
\item{archive}{A \code{\link[flapper]{pf_archive-class}} object.}

\item{bathy}{(optional) A \code{\link[raster]{raster}} that defines the grid across the area over which particle filtering was applied. If unsupplied, this is extracted from \code{archive} if available.}
}
\value{
The function returns a dataframe that defines, for each time step (`timestep'), particle samples (`cell_id'), associated probabilities (`cell_pr') and, if \code{bathy} is available, cell coordinates (`cell_x', `cell_y' and `cell_z').
}
\description{
This function accesses and simplifies the `history' list in a \code{\link[flapper]{pf_archive-class}} object.
}
\details{
From the `history' element of a \code{\link[flapper]{pf_archive-class}} object, this function extracts particle samples as a dataframe with columns for time steps, cell IDs, cell probabilities and coordinates (if \code{bathy} is available).
}
\examples{
pf_access_history(dat_dcpf_histories)
}
\author{
Edward Lavender
}
