% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_support.R
\name{pf_setup_movement_pr}
\alias{pf_setup_movement_pr}
\title{A simple movement model dependent on distance}
\usage{
pf_setup_movement_pr(distance, ...)
}
\arguments{
\item{distance}{A numeric vector of distances (m).}

\item{...}{Additional arguments (none implemented).}
}
\value{
The function returns a numeric vector of probabilities that represent the probability of movement between two or more areas given the distances between them.
}
\description{
This function provides a simple movement model that calculates the probability of movement between two locations according to the distance between them, using an logistic equation with pre-defined parameters.
}
\details{
Under this model, for distance(s) \eqn{ \leq 500 } m, \eqn{Pr(distance) = logistic(10 + distance  -0.05)}; otherwise, \eqn{Pr(distance) = 0}. This particular model is designed for flapper skate (\emph{Dipturus intermedius}) and represents a reasonable model for the probability of moving a given distance in a two-minute period (in the absence of additional information).
}
\examples{
pr <- pf_setup_movement_pr(1:1000)
plot(pr, type = "l", xlab = "Distance (m)", ylab = "Pr(distance)")
}
\seealso{
This function is used as the default movement model in \code{\link[flapper]{pf}}.
}
\author{
Edward Lavender
}
