% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_similarity_matrices.R
\name{plot_det_sim}
\alias{plot_det_sim}
\title{Plot a detection similarity matrix}
\usage{
plot_det_sim(
  z,
  zlim = NULL,
  col = (grDevices::colorRampPalette(c("white", "yellow", "orange", "red")))(100),
  col_diag = NULL,
  grid = NULL,
  cex.axis = 1,
  xlab = "Individual",
  ylab = "Individual",
  ...
)
}
\arguments{
\item{z}{A (detection similarity) matrix to be plotted (see \code{\link[flapper]{compute_det_sim}}). Individual IDs (for the axis tick mark labels) are extracted from the row names of this matrix, if provided. Otherwise, axis tick mark labels are given as an index.}

\item{zlim}{A numeric vector of length two which specifies the z axis limits (see \code{\link[fields]{image.plot}}),}

\item{col}{A colour table to use for the image (see \code{\link[fields]{image.plot}}).}

\item{col_diag}{(optional) A colour which, if provided, is used to shade the diagonal of the matrix.}

\item{grid}{(optional) A named list which, if provided, is passed to \code{\link[plot.pretty]{add_grid_rect_xy}} to draw grid lines around each matrix cell. Arguments \code{x} and \code{y} default to the positions of each matrix cell and do not need to be provided. To use \code{\link[plot.pretty]{add_grid_rect_xy}}'s default graphical options, simply specify \code{grid = list()}.}

\item{cex.axis}{A number which specifies the axis font size (specifically, the magnification to be used for axis annotation relative to the current setting of \code{cex}, see \code{\link[graphics]{par}}).}

\item{xlab}{The x axis label.}

\item{ylab}{The y axis label.}

\item{...}{Additional arguments passed to \code{\link[fields]{image.plot}}. These should not include \code{x}, \code{y}, \code{xlim}, \code{ylim} or \code{axes} which are controlled internally.}
}
\value{
The function returns a plot of a similarity matrix.
}
\description{
This function is a wrapper for \code{\link[fields]{image.plot}} designed to streamline the visualisation of detection similarity matrices. The function returns a plot which can be customised.
}
\examples{

}
\author{
Edward Lavender
}
