% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number}
\source{
This function and the documentation are taken from the `plyr' package. The function is defined separately in \code{\link[flapper]{flapper}} to reduce reliance on non-default packages.
}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{numeric or date-time (POSIXct) vector to round}

\item{accuracy}{number to round to; for POSIXct objects, a number of seconds}

\item{f}{rounding function: floor, ceiling or round}
}
\description{
Round to multiple of any number.
}
\keyword{internal}
