% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{sim_path_*}
\alias{sim_path_*}
\title{Functions for the simulation of movement paths}
\description{
\link{flapper} includes a number of functions for the simulation of movement paths (\code{sim_path_*()}), listed here for convenience:
\itemize{
  \item{\link{sim_path_sa}} - simulate discrete-time movement paths from step lengths and turning angles
  \item{\link{sim_path_ou_1}} -- simulate discrete-time movement paths under a Ornstein-Uhlenbeck process with time-fixed parameters
}
}
