% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tools.R
\name{update_extent}
\alias{update_extent}
\title{Shrink or expand an \code{\link[raster]{extent}} object}
\usage{
update_extent(x, x_shift, y_shift = x_shift)
}
\arguments{
\item{x}{A \code{\link[raster]{raster}} or an \code{\link[raster]{extent}} object.}

\item{x_shift}{A number that defines the change in the x limits. Positive numbers increase the extent (i.e., the lower limit is reduced and the upper limit is increased) while negative numbers reduce extent (i.e., the lower limit is increased and the upper limit is decreased).}

\item{y_shift}{A number that defines the change in the y limits. By default, this equals \code{x_shift} (i.e., the \code{\link[raster]{extent}} is changed uniformly in all directions).}
}
\value{
The function returns an updated \code{\link[raster]{extent}} object.
}
\description{
This function updates a \code{\link[raster]{raster}}'s \code{\link[raster]{extent}} by shrinking or expanding the x and y limits.
}
\examples{
# Define example raster
r <- raster::raster(nrows = 3, ncols = 3,
                    resolution = c(5, 5),
                    ext = raster::extent(0, 15, 0, 15))
# Increase raster extent by 5 units in every direction
update_extent(r, 5)
# Decrease raster extent by 5 units in every direction
update_extent(r, -5)
# Increase x and y extent parameters differently
update_extent(r, 5, 10)

}
\author{
Edward Lavender
}
