% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc_classes.R
\docType{package}
\name{acdc-class}
\alias{acdc-class}
\title{"acdc" class}
\value{
An \code{\link[flapper]{acdc-class}} object is a named list that contains the output of a call to \code{\link[flapper]{acdc}}. This contains the following elements: `.acdc', `ts_by_chunk', `time' and `args'.

The main output is the `.acdc' element. This contains a list of arguments returned by the call(s) to the \code{\link[flapper]{.acdc}} function, which is the workhorse that actually implements the algorithm. If the algorithm is implemented step-wise, this contains a single \code{\link[flapper]{.acdc-class}} object as returned by \code{\link[flapper]{.acdc}}. If the algorithm is implemented chunk-wise, this is a list, with one element for each chunk, each element containing a \code{\link[flapper]{.acdc-class}} object with results of the call to \code{\link[flapper]{.acdc}} for that chunk. The results across chunks can be aggregated using \code{\link[flapper]{acdc_simplify}}.

The `ts_by_chunk' element is a list, with one element for each chunk, that contains the acoustic and archival time series that were used in that chunk. If there is more than one chunk, the last observation of each acoustic chunk is the same as the first acoustic observation for the next chunk. This results from splitting chunks at acoustic observations, which enables the results of chunks that are computed independently to be simply aggregated without the loss of any information.

The `time' element is a dataframe that defines the times of sequential stages in the algorithm's progression, providing a record of computation time.

The `args' element is a named list of user inputs that records the parameters used to generate the outputs.
}
\description{
An S3 class that defines the object returned by the \code{\link[flapper]{acdc}} algorithm.
}
\author{
Edward Lavender
}
