% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdc.R
\name{acdc_simplify}
\alias{acdc_simplify}
\title{Simplify the outputs of \code{\link[flapper]{acdc}}}
\usage{
acdc_simplify(acdc, keep_chunks = FALSE, ...)
}
\arguments{
\item{acdc}{A \code{\link[flapper]{acdc-class}} object returned by \code{\link[flapper]{acdc}}.}

\item{keep_chunks}{A logical variable that defines whether or not to retain all chunk-specific information.}

\item{...}{Additional arguments (none implemented).}
}
\value{
The function returns an object of class \code{\link[flapper]{.acdc-class}}.
}
\description{
This function simplifies the output of \code{\link[flapper]{acdc}}, by extracting elements from the '.acdc' elements which hold the results of calls to the workhorse function \code{\link[flapper]{.acdc}}. This is especially useful if the ACDC algorithm has been applied chunk-wise, in which case the results for each acoustic chunk are returned in a list. This function aggregates information across chunks to generate a continuous time series of results and a map of where the individual could have spend more or less time over the entire time series.
}
\details{
If the \code{\link[flapper]{acdc}} function was implemented step-wise, this function simply extracts the information into the format of an \code{\link[flapper]{.acdc-class}} object. For a chunk-wise implementation, the function (a) computes the map of where the individual could have spent more or less time by aggregating the chunk-specific maps, accounting for the overlap between chunks; and (b) simplifies chunk-specific records into a single contiguous time series and with re-defined time stamps from the start to the end of the time series and then returns an \code{\link[flapper]{.acdc-class}} object.
}
\author{
Edward Lavender
}
