% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check...}
\alias{check...}
\title{Check that arguments supplied via ... are allowed}
\usage{
check...(not_allowed, ...)
}
\arguments{
\item{not_allowed}{A character vector of the names of function arguments that are not allowed.}

\item{...}{Other arguments}
}
\value{
The function checks other arguments supplied via ...; if these contain an argument that is not allowed, the function returns an error. Otherwise, nothing is returned.
}
\description{
This function checks that arguments supplied via ... are allowed. This function was written to support other functions, specifically via the return of a helpful error message if arguments that cannot be supplied via ... have been supplied. The function is not intended for general use.
}
\author{
Edward Lavender
}
\keyword{internal}
