% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_acdc}
\alias{dat_acdc}
\title{Example ACDC algorithm output}
\format{
A named list with 4 elements:
\describe{
  \item{map}{A RasterLayer that shows where the individual could have spent more or less time over the duration of the movement time-series.}
  \item{record}{A list that records time-specific maps of the possible locations of the individual, and can be used to plot maps of specific time points or to produce animations.}
  \item{time}{A dataframe that defines the times of sequential stages in the algorithm's progression.}
  \item{args}{A named list of user inputs that record the parameters used to generate the outputs.}
}
}
\usage{
dat_acdc
}
\description{
A list of outputs created by \code{\link[flapper]{.acdc}}. This is included principally to streamline function examples.
}
\seealso{
Further details are given in \code{\link[flapper]{.acdc}}.
}
\keyword{datasets}
