% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_acoustics}
\alias{dat_acoustics}
\title{Example passive acoustic telemetry detections dataset}
\format{
A dataframe with 39,242 observations and 9 variables:
\describe{
  \item{individual_id}{A unique identifier of the individual that was detected (see \code{\link{dat_ids}}).}
  \item{transmitter_id}{The acoustic transmitter that was detected (see \code{\link{dat_ids}}).}
  \item{index}{An integer individual-specific cumulative detection index.}
  \item{timestamp}{A POSIXct object which defines the time of each observation.}
  \item{receiver_id}{A unique identifier of the receiver at which the individual was detected (see \code{\link{dat_moorings}}).}
  \item{receiver}{The receiver at which the individual was detected (see \code{\link{dat_moorings}}).}
  \item{receiver_long}{The longitude of the receiver at which the individual was detected (see \code{\link{dat_moorings}}).}
  \item{receiver_lat}{The latitude of the receiver at which the individual was detected (see \code{\link{dat_moorings}}).}
  \item{receiver_depth}{The depth of the receiver at which the individual was detected (see \code{\link{dat_moorings}}).}
}
}
\source{
Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.
}
\usage{
dat_acoustics
}
\description{
A dataset containing a sample of processed flapper skate (\emph{Dipturus intermedius}) detection time series. Data are arranged by \code{individual_id}, \code{timestamp} and then \code{receiver_id} (see below).
}
\details{
Time stamps are expressed to the nearest two minutes. `Duplicate' detections (of the same individual at the same receiver in the same time step) are excluded. This supports the implementation of the acoustic-container (AC) and acoustic-container depth-contour (ACDC) algorithms (see \code{\link[flapper]{ac}} and \code{\link[flapper]{acdc}}).
}
\references{
Data collection and processing is described in Lavender (in prep). Fine-scale habitat use of the Critically Endangered flapper skate (\emph{Dipturus intermedius}). [Doctoral dissertation, University of St Andrews].
}
\keyword{datasets}
