% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_dc}
\alias{dat_dc}
\title{Example DC algorithm output}
\format{
A named list with 4 elements:
\describe{
  \item{dc}{A \code{\link[raster]{raster}} that defines the number of times that the depth in each cell overlapped with the depth recorded in the individual's \code{archival} time series.}
  \item{archival}{The \code{archival} dataframe inputted to \code{\link[flapper]{dc}}. This includes two new columns that define the lower and upper bounds for the possible depth of the individual on the bathymetry raster (\code{bathy}) at each time step (`depth_lwr' and `depth_upper') derived from \code{calc_depth_error} and, if \code{calc_availability = TRUE}, a logical vector that defines whether or not there are any cells on \code{bathy} of the required depth range at each time step.}
  \item{args}{A named list of user inputs that record the parameters used to generate the outputs.}
}
}
\usage{
dat_dc
}
\description{
A named list of outputs from \code{\link[flapper]{dc}}. This is included principally to streamline function examples.
}
\note{
\code{\link[flapper]{dc}} outputs may also contain a `spatial' element. This has not been retained for \code{\link[flapper]{dat_dc}} to minimise dataset size.
}
\seealso{
See \code{\link[flapper]{dc}} for further information.
}
\keyword{datasets}
