% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_ids}
\alias{dat_ids}
\title{Example tagged individuals dataset}
\format{
A dataframe with 33 observations and 6 variables:
\describe{
  \item{individual_id}{A number which distinguishes each unique individual.}
  \item{transmitter_id}{A character which defines the acoustic transmitter code.}
  \item{archival_id}{A character which defines the archival tag code.}
  \item{tag_start_date}{A date which defines the date of tag deployment.}
  \item{tag_long}{A number which defines the longitude (decimal degrees) of tag deployment.}
  \item{tag_lat}{A number which defines the latitude (decimal degrees) of tag deployment.}
}
}
\source{
Data were collected by, and belong to, Marine Scotland Science and NatureScot. Data were processed by Edward Lavender. If you wish to use these data, please contact Marine Scotland Science and NatureScot for further information.
}
\usage{
dat_ids
}
\description{
A dataset containing the characteristics of a sample of flapper skate (\emph{Dipturus intermedius}) that were tagged as part of the Movement Ecology of Flapper Skate (MEFS) project. Data are arranged by \code{individual_id} (see below).
}
\references{
Data collection and processing are described in Lavender (2022). Modelling the movements of flapper skate (\emph{Dipturus intermedius}) in relation to a Scottish Marine Protected Area. University of St Andrews. https://www.doi.org/10.17630/sta/201
}
\keyword{datasets}
