% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dats.R
\docType{data}
\name{dat_moorings}
\alias{dat_moorings}
\title{Example passive acoustic telemetry receiver moorings dataset}
\format{A dataframe with 39 observations and 8 variables:
\describe{
  \item{receiver_id}{A number which distinguishes each unique receiver deployment.}
  \item{receiver}{A character which defines each receiver.}
  \item{sentinel_id}{A character which defines each receiver's built-in acoustic transmitter ('sentinel tag'). Some receivers did not have sentinel tags; their \code{sentinel_id} is \code{NA}.}
  \item{receiver_start_date}{A date which defines the start date of the receiver's deployment.}
  \item{receiver_end_date}{A date which defines the end date of the receiver's deployment.}
  \item{receiver_long}{A number which defines the latitude (decimal degrees) of the receiver.}
  \item{receiver_lat}{A number which defines the latitude (decimal degrees) of the receiver.}
  \item{receiver_depth}{A number which defines the approximate depth of each receiver below the surface (m).}
}}
\source{
Data were collected by Marine Scotland Science and Scottish Natural Heritage. Data were processed by Edward Lavender.
}
\usage{
dat_moorings
}
\description{
A dataset containing a sample of passive acoustic telemetry receiver locations and associated information. Data are arranged by \code{receiver_id} (see below).
}
\keyword{datasets}
