% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpf.R
\name{dcpf_simplify}
\alias{dcpf_simplify}
\title{Convert particle histories from \code{\link[flapper]{dcpf}} into movement paths}
\usage{
dcpf_simplify(record, verbose = TRUE)
}
\arguments{
\item{record}{A \code{\link[flapper]{.dcpf-class}} object from \code{\link[flapper]{dcpf}}.}

\item{verbose}{A logical input that defines whether or not to print messages to the console to monitor function progress.}
}
\value{
The function returns a \code{\link[flapper]{dcpf-class}} object, which is a dataframe that defines the movement paths.
}
\description{
This function is designed to simplify the \code{\link[flapper]{.dcpf-class}} object from \code{\link[flapper]{dcpf}} that defines sampled particle histories into a set of movement paths.
}
\details{
If there are vast numbers of possible paths through the landscape captured by particle movements, the function can run into vector memory limitations when assembling paths.
}
\examples{
# An example with the example DCPF particle histories dataset included in flapper
paths <- dcpf_simplify(dat_dcpf_histories)

}
\author{
Edward Lavender
}
