% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dist_btw_receivers}
\alias{dist_btw_receivers}
\title{Compute Euclidean distances between receivers}
\usage{
dist_btw_receivers(moorings, f = NULL)
}
\arguments{
\item{moorings}{A dataframe which defines each unique receiver deployment. This should contain the columns: `receiver_id', a unique identifier of each receiver, `receiver_lat', the latitude of that receiver in decimal degrees; and `receiver_long', the longitude of that receiver in decimal degrees (see \code{\link[flapper]{dat_moorings}}).}

\item{f}{(optional) A function which is used to process distances before these are returned. For example, it may be useful to round distances to nearest km with \code{f = function(x) round(x, digits = 0)}.}
}
\value{
The function returns a dataframe with columns `r1', `r2' and `dist'. These define the IDs of each combination of receivers and the associated distance between them, in km. Note that the dataframe contains duplicate combinations of receivers (e.g., both r1 = 1 and r2 = 2 and r1 = 2 and r2 = 1).
}
\description{
This function computes Euclidean distances (km) between all combinations of receivers.
}
\examples{
#### Example (1): Compute distances between all combinations of receivers in km
# Define dataframe with required columns
dat <- data.frame(receiver_id = dat_moorings$receiver_id,
                  receiver_lat = dat_moorings$receiver_lat,
                  receiver_long = dat_moorings$receiver_long)
# Compute distances
dist_btw_receivers_km <- dist_btw_receivers(dat)
head(dist_btw_receivers_km)

#### Example (2): Post-process distances via the f argument
dist_btw_receivers_km_round <- dist_btw_receivers(dat, f = round)
head(dist_btw_receivers_km_round)
# convert distances to m
dist_btw_receivers_m <- dist_btw_receivers(dat, f = function(x) x*1000)
head(dist_btw_receivers_m)

}
\author{
Edward Lavender
}
